+++ tests/test_formatting.py
@@ -1,6 +1,6 @@
 """Tests for `ember_mug.formatting`."""
 from ember_mug.data import Colour
-from ember_mug.formatting import format_led_colour, format_liquid_level, format_temp
+from ember_mug.formatting import format_capacity, format_led_colour, format_liquid_level, format_temp
 
 
 def test_format_led_colour() -> None:
@@ -15,3 +15,9 @@ def test_format_liquid_level() -> None:
 def test_format_temp() -> None:
     assert format_temp(25.445) == "25.45°C"
     assert format_temp(36.443, metric=False) == "36.44°F"
+
+
+def test_format_capacity() -> None:
+    assert format_capacity(None) == "Unknown"
+    assert format_capacity(414) == "414ml"
+    assert format_capacity(414, metric=False) == "14oz"